flatcode/fatcode_sample/requirements.txt
flask
Flask-HTTPAuth
pyyaml
rsa